feat: abstract key store handler for various indexing logic
