+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/model/IndexBasedKeyStoreHandlerProperty.kt
@@ -8,19 +8,23 @@ import javax.annotation.PostConstruct
 @ConfigurationProperties(prefix = "graphene.writer.store.key.handlers.index-based-key-store-handler")
 data class IndexBasedKeyStoreHandlerProperty(
   var enabled: Boolean = false,
-  var clusterName: String? = null,
-  var index: String? = null,
-  var type: String? = null,
   var isCache: Boolean = false,
   var expire: Long = 0,
-  var cluster: List<String> = mutableListOf(),
-  var port: Int = 0,
-  var bulk: IndexBulkConfiguration? = null
-) {
+  override var clusterName: String = "graphene",
+  override var index: String = "metric",
+  override var type: String = "path",
+  override var cluster: List<String> = mutableListOf(),
+  override var port: Int = 0,
+  override var bulk: IndexBulkConfiguration? = null
+) : ElasticsearchKeyStoreHandlerProperty(clusterName, index, type, cluster, port, bulk) {
 
   @PostConstruct
   fun init() {
     logger.info("Load Graphene IndexBasedKeyStoreHandlerProperty : {}", toString())
+  }
+
+  override fun toString(): String {
+    return "IndexBasedKeyStoreHandlerProperty(enabled=$enabled, isCache=$isCache, expire=$expire, clusterName='$clusterName', index='$index', type='$type', cluster=$cluster, port=$port, bulk=$bulk)"
   }
 
   companion object {
