+++ graphene-writer/src/main/kotlin/com/graphene/writer/store/key/model/SimpleKeyStoreHandlerProperty.kt
@@ -6,21 +6,25 @@ import org.springframework.boot.context.properties.ConfigurationProperties
 import javax.annotation.PostConstruct
 
 @ConfigurationProperties(prefix = "graphene.writer.store.key.handlers.simple-key-store-handler")
-data class SimpleKeyStoreHandlerProperty(
+class SimpleKeyStoreHandlerProperty(
   var enabled: Boolean = false,
-  var clusterName: String? = null,
-  var index: String? = null,
-  var type: String? = null,
   var isCache: Boolean = false,
   var expire: Long = 0,
-  var cluster: List<String> = mutableListOf(),
-  var port: Int = 0,
-  var bulk: IndexBulkConfiguration? = null
-) {
+  override var clusterName: String = "graphene",
+  override var index: String = "metric",
+  override var type: String = "path",
+  override var cluster: List<String> = mutableListOf(),
+  override var port: Int = 9300,
+  override var bulk: IndexBulkConfiguration? = null
+) : ElasticsearchKeyStoreHandlerProperty(clusterName, index, type, cluster, port, bulk) {
 
   @PostConstruct
   fun init() {
     logger.info("Load Graphene SimpleKeyStoreHandlerProperty : {}", toString())
+  }
+
+  override fun toString(): String {
+    return "SimpleKeyStoreHandlerProperty(enabled=$enabled, isCache=$isCache, expire=$expire, clusterName='$clusterName', index='$index', type='$type', cluster=$cluster, port=$port, bulk=$bulk)"
   }
 
   companion object {
